helloagents-trip-planner/backend/app/config.py
"""配置管理模块"""

import os
from pathlib import Path
from typing import List
from pydantic import Field
from pydantic_settings import BaseSettings
from dotenv import load_dotenv

# 加载环境变量
# 首先尝试加载当前目录的.env
load_dotenv()

# 当前文件位置：backend/config.py
env_path = Path(__file__).parent.parent / ".env"

if env_path.exists():
    load_dotenv(env_path, override=False)  # 不覆盖已有的环境变量


class Settings(BaseSettings):
    """应用配置"""

    # 应用基本配置
    app_name: str = "HelloAgents智能旅行助手"
    app_version: str = "1.0.0"
    debug: bool = False

    # 服务器配置
    host: str = "0.0.0.0"
    port: int = 8000

    # CORS配置 - 使用字符串,在代码中分割
    cors_origins: str = "http://localhost:5173,http://localhost:3000,http://127.0.0.1:5173,http://127.0.0.1:3000"

    # 高德地图API配置
    amap_api_key: str = ""  # 保持向后兼容（可选）
    amap_app_code: str = ""  # 高德地图Web服务API密钥

    # Unsplash API配置
    unsplash_access_key: str = ""
    unsplash_secret_key: str = ""

    # LLM配置 (从环境变量读取,由HelloAgents管理)
    llm_api_key: str = ""
    llm_base_url: str = ""
    llm_model: str = ""

    # ================= Redis缓存配置 =================
    # Redis连接配置
    redis_host: str = ""
    redis_port: int = 6379
    redis_db: int = 0
    redis_password: str = ""

    # Redis前缀（用于区分不同应用的缓存key）
    redis_prefix: str = "trip_planner"

    # 缓存总开关
    redis_enabled: bool = True

    # ================= 数据库持久化配置 =================
    # 使用 TRIP_DB_URL 避免与系统 DATABASE_URL 环境变量冲突
    database_url: str = Field(
        default="sqlite+aiosqlite:///./data/trip_planner.db",
        alias="TRIP_DB_URL",
    )
    database_auto_create: bool = True

    # 日志配置
    log_level: str = "INFO"

    class Config:
        env_file = ".env"
        case_sensitive = False  # 禁用大小写敏感匹配
        extra = "ignore"  # 忽略额外的环境变量

    def get_cors_origins_list(self) -> List[str]:
        """获取CORS origins列表"""
        return [origin.strip() for origin in self.cors_origins.split(',')]


# 创建全局配置实例
settings = Settings()


def get_settings() -> Settings:
    """获取配置实例"""
    return settings


# 验证必要的配置
def validate_config():
    """验证配置是否完整"""
    errors = []
    warnings = []

    if not settings.amap_app_code:
        errors.append("AMAP_APP_CODE未配置（必需，高德地图Web服务API密钥）")
    elif not settings.amap_api_key:
        warnings.append("AMAP_API_KEY未配置，建议同时配置")

    # LLM会自动从LLM_API_KEY读取
    llm_api_key = os.getenv("LLM_API_KEY")
    if not llm_api_key:
        warnings.append("LLM_API_KEY未配置,LLM功能可能无法使用")

    if errors:
        error_msg = "配置错误:\n" + "\n".join(f"  - {e}" for e in errors)
        raise ValueError(error_msg)

    if warnings:
        print("\n⚠️  配置警告:")
        for w in warnings:
            print(f"  - {w}")

    return True


# 打印配置信息(用于调试)
def print_config():
    """打印当前配置(隐藏敏感信息)"""
    print(f"应用名称: {settings.app_name}")
    print(f"版本: {settings.app_version}")
    print(f"服务器: {settings.host}:{settings.port}")
    print(f"高德地图API Key: {'已配置' if settings.amap_api_key else '未配置'}（可选）")
    print(f"高德地图App Code: {'已配置' if settings.amap_app_code else '未配置'}（必需）")

    # 检查LLM配置
    llm_api_key = os.getenv("LLM_API_KEY")
    llm_base_url = os.getenv("LLM_BASE_URL") or settings.llm_base_url
    llm_model = os.getenv("LLM_MODEL") or settings.llm_model

    print(f"LLM API Key: {'已配置' if llm_api_key else '未配置'}")
    print(f"LLM Base URL: {llm_base_url}")
    print(f"LLM Model: {llm_model}")
    print(f"日志级别: {settings.log_level}")

