helloagents-trip-planner/backend/requirements.txt
# LangChain框架
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langgraph>=0.2.0

# FastAPI和相关依赖
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP客户端
httpx>=0.27.0
aiohttp>=3.10.0

# 数据库持久化
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0

# 环境变量管理
python-dotenv>=1.0.0

# CORS支持
python-multipart>=0.0.9

# 日志
loguru>=0.7.0

# 其他工具
python-dateutil>=2.8.2
huggingface_hub>=0.25.0

# Redis缓存层
redis>=5.0.0
